third_party/WebKit/LayoutTests/inspector/sources/debugger/promise-tracker.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function testFunction()
{
    var promise = new Promise(function promiseConstructor(resolve, reject) {
        resolve("Resolved!");
    });
    promise
        .then(thenCallback, errorCallback)
        .then(thenCallback2, errorCallback);
}

function thenCallback() { }

function thenCallback2()
{
    debugger;
}

function errorCallback() { }

var test = function ()
{
    InspectorTest.startDebuggerTest(step1);

    function step1()
    {
        DebuggerAgent.enablePromiseTracker();
        InspectorTest.runTestFunctionAndWaitUntilPaused(step2);
    }

    function step2()
    {
        DebuggerAgent.getPromises(didGetPromises);
    }

    function didGetPromises(error, response)
    {
        InspectorTest.assertTrue(!error, "FAIL: " + error);

        function comparePromiseData(x, y)
        {
            if (x.id < y.id)
                return -1;
            else if (x.id === y.id)
                return 0;
            else
                return 1;
        }

        var minPromiseId;
        if (response.length) {
            minPromiseId = response[0].id;
            for (var i = 0; i < response.length; i++)
                minPromiseId = Math.min(minPromiseId, response[i].id);
        }
        response.sort(comparePromiseData);

        var output = [];
        for (var i = 0; i < response.length; i++) {
            var promise = response[i];
            var parentId = promise.parentId ? promise.parentId - minPromiseId : undefined;
            var promiseInfo = "Promise:" +
                              "\n    id: " + (promise.id - minPromiseId) +
                              "\n    status: " + promise.status +
                              "\n    parent id: " + parentId;
            var callFrame = promise.callFrame;
            if (callFrame) {
                var url = WebInspector.displayNameForURL(callFrame.url);
                promiseInfo += "\n    " + callFrame.functionName + " " + url + ":" + callFrame.lineNumber;
            }
            output.push(promiseInfo);
        }

        InspectorTest.addResults(output);

        DebuggerAgent.getPromiseById(response[0].id, "console", didGetPromiseById);
    }

    function didGetPromiseById(error, response)
    {
        InspectorTest.assertTrue(!error, "FAIL: " + error);
        InspectorTest.assertTrue(response && response.type === "object" && response.className === "Promise", "FAIL: returned by ID object is not a Promise.");

        RuntimeAgent.getProperties(response.objectId, didGetProperties);
    }

    function didGetProperties(error, properties, internalProperties) {
        InspectorTest.assertTrue(!error, "FAIL: " + error);
        InspectorTest.assertTrue(internalProperties && internalProperties.length > 0, "FAIL: no internal properties.");

        var status, value;
        for (var i = 0; i < internalProperties.length; i++) {
            var property = internalProperties[i];
            if (property.name === "[[PromiseStatus]]")
                status = property.value.value;
            if (property.name === "[[PromiseValue]]")
                value = property.value.value;
        }

        InspectorTest.addResult("Promise:\n    status: " + status + "\n    value: " + value);

        DebuggerAgent.disablePromiseTracker();
        DebuggerAgent.getPromises(didGetPromises2);
    }

    function didGetPromises2(error, response)
    {
        var hasData = !error && response && response.length > 0;
        InspectorTest.assertTrue(!hasData, "FAIL: expected error or no data after disabling promise tracker.");
        InspectorTest.completeDebuggerTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests promise tracker in debugger.
</p>
</body>
</html>
