third_party/WebKit/LayoutTests/inspector/sources/debugger/promise-tracker-expected.txt

Tests promise tracker in debugger.

Set timer for test function.
Script execution paused.
Promise:
    id: 0
    status: resolved
    parent id: undefined
    testFunction promise-tracker.html:9
Promise:
    id: 1
    status: resolved
    parent id: 0
    testFunction promise-tracker.html:13
Promise:
    id: 2
    status: pending
    parent id: 1
    testFunction promise-tracker.html:14
Promise:
    status: resolved
    value: Resolved!
Script execution resumed.
